==> 03-Evaluation/2023/00-createSummaries/04-MergeCollapsed.ipynb <==
import os
import json
import pandas as pd
import numpy as np
from itertools import product

from sentence_transformers import SentenceTransformer, util

OLD_FOLDER = './collapsed-event-days-old'
NEW_FOLDER = './collapsed-event-days-new'
SAVE_FOLDER = './collapsed-event-days'

embedding_model = SentenceTransformer("BAAI/bge-m3", device="cuda")

for file in os.listdir(OLD_FOLDER):
    if file.endswith('.json'):
        new_file = os.path.join(NEW_FOLDER, file)
        old_file = os.path.join(OLD_FOLDER, file)
        save_path = os.path.join(SAVE_FOLDER, file)

        # Load JSON data
        with open(old_file, 'r') as f:
            old_data = json.load(f)
        with open(new_file, 'r') as f:
            new_data = json.load(f)

        # Convert old_data to a dictionary for faster lookup
        old_data_dict = {entry["collapsed_fact_id"]: entry for entry in old_data}

        for new_entry in new_data:
            fact_id = new_entry["collapsed_fact_id"]
            if fact_id in old_data_dict:
                old_entry = old_data_dict[fact_id]
                if new_entry["fact_text"] not in old_entry["fact_text"]:
                    old_entry["fact_text"] = " ".join([old_entry["fact_text"], new_entry["fact_text"]])
                    print("Merging facts")
                else:
                    old_embedding = embedding_model.encode(old_entry["fact_text"], convert_to_tensor=True)
                    new_embedding = embedding_model.encode(new_entry["fact_text"], convert_to_tensor=True)
                    similarity = util.pytorch_cos_sim(old_embedding, new_embedding).item()
                    if similarity < 0.96:
                        old_entry["fact_text"] = " ".join([old_entry["fact_text"], new_entry["fact_text"]])
                old_entry["num_relevant_facts"] += new_entry["num_relevant_facts"]
                old_entry["importance"] = max(old_entry["importance"], new_entry["importance"])
                old_entry["relevant_facts"] = list(set(old_entry["relevant_facts"] + new_entry["relevant_facts"]))
            else:
                old_data.append(new_entry)  # If new fact ID is not in old data, add it

        # Save merged JSON
        with open(save_path, 'w') as f:
            json.dump(old_data, f, indent=4)
